refactor: move isOperationParameterRequired from parser to parser utils
