+++ packages/openapi-ts/src/generate/types.ts
@@ -8,9 +8,9 @@ import {
 } from '../compiler';
 import type { Model, OperationParameter } from '../generator/openApi';
 import type { Method } from '../generator/openApi';
-import { isOperationParameterRequired } from '../generator/openApi/common/parser/operation';
 import { enumEntry, enumUnionType } from '../generator/utils/enum';
 import { escapeComment } from '../generator/utils/escape';
+import { isOperationParameterRequired } from '../generator/utils/operation';
 import type { Client } from '../types/client';
 import type { Files } from '../types/utils';
 import { getConfig, isStandaloneClient } from '../utils/config';
